CREDENTIAL: Restructure 'get()'.
Based on the conversation in w3c/webappsec#292, and the critique in
w3c/webappsec#349, this patch restructures `get()` to accept two
arguments: a sequence of types, and a dictionary of dictionaries
that specifies the filter options for each relevant type.

This does not completely address the request in #292, as the matching
algorithm is not exposed as a method on LocallyStoredCredential. I
need to think a bit more about the impact of doing that. Splitting up
the patch to make sure the pieces we know we need get in for feedback.