Fix easy bikeshed warnings for index.html
Bikeshed complains that the variables inputBody, memory, undefined, init and
deep clone are only used once. The reasons are:
 * undefined should actualy be a <code> not a <var> element, because it
   expresses a JavaSCript constant, not a variable
 * all the rest references external algorithms, which define them

This patch marks all of those <var>s with the "ignore" attribute to silence the
warning.

Note that there are two more warnings for the unique mention: for |possible
types| and |Content-Type|. The former is in fact a malformed algorithm, and the
second seems like an error with bikeshed, because it has multiple occurrences.
Those two will be dealt with separately.