Clarify the "Request a Credential" algorithm
This CL drops the input |types| from the "Request a Credential" algorithm,
because it is not used (instead, |types| is a local variable, value of which is
computed by the algorithm).

The CL also replaces "name" of an interface with a [[type]] shared by instances
implementing this interface. For PasswordCredential, for instance, "name" is
PasswordCredential, whereas [[type]] is "password". The latter corresponds to a
possible key in CredentialRequestOptions, which is what the algorithm expects.

This should fix issue #41.