Remove the blanket restriction against nested usage.
Some credential types wish to be usable in nested contexts. This patch moves the
top-level restriction out of the 'navigator.credentials' API itself, and into the
credential types that wish to have such a restriction.

Note that it also changes the behavior of those restrictions: rather than throwing
a 'NotSupported' error, it instead returns an empty set. That seems both simpler and
less likely to cause developers to be angry at me.