Update README.md
Hello!
First, let me thank you for make this dependency available. It was very helpful for me :)

I would suggest you to update your README.md file for clarify to possibles angular X users how to use this dependency because it was not evident for me to find the way to do it. I wasted two hours to find the response to my problems (I know, i'm a noob :) But i'm pretty sure that can be helpful to some people! )

Please, if you want use different words or put a more rich text, be free to do it or let me know if I can change something or put this suggestion in another location! (probably it would be nice to have this explanation on 'Details' section of the npmjs web site : https://www.npmjs.com/package/@types/webappsec-credential-management )

Thanks in advance for your response.

King regards,
Eduardo Borges Da Silva